refactor(backend): split master_catalog.go by actor and extract the repeated published gate
